+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### C# -> VB
 * XML output path needs output path prepending [#641](https://github.com/icsharpcode/CodeConverter/issues/641)
 
+* Converting explicit and implicit operator failed [#659](https://github.com/icsharpcode/CodeConverter/issues/659)
 
 ## [8.2.0] - 2020-10-12